feature - all darktable-cli working; changes to the export structure
Changed the export structure from An Export class that you define at the start and holds the cache intance and all export related options from which you call export method on photo
to: export being a function you call on each photo with its own parameters each time. and added the cache instance as a separate class.
Most underlying code is the same
But i belive this way is more flexible per export (if the export arguments need to change for each export you can just do it intead of creating a new Export instance for each combination of arguments)
Seems to me the only reason to have an Export class was to hold the cache intance (which you may not need anyway if you just want to use exprt() instead of export_cached() but this way also forced you too create one) so lets move the cache into its own class and pass that to export_cached()
I really think this makes more sense; so much so that i broke backwards compatibility
also export_cached() has been renamed to export_with_cache() which i think it slightly conveys better what it does (since i was already breaking backwards compatibility...)
